fix いろいろ
