+++ resources/assets/js/checkin.tsx
@@ -3,4 +3,4 @@ import * as ReactDOM from 'react-dom';
 import { CheckinForm } from './components/CheckinForm';
 
 const initialState = JSON.parse(document.getElementById('initialState')?.textContent as string);
-ReactDOM.render(<CheckinForm initialState={initialState}/>, document.getElementById('checkinForm'));
+ReactDOM.render(<CheckinForm initialState={initialState} />, document.getElementById('checkinForm'));
